Refactor Repository & Service for architecture
