+++ src/main/java/com/example/utmentor/models/webModels/PageResponse.java
@@ -9,4 +9,15 @@ public record PageResponse<T>(
     long total,
     int totalPages,
     boolean hasNext
-) {}
+) {
+    public PageResponse(List<T> data, int page, int pageSize, long total) {
+        this(
+            data,
+            page,
+            pageSize,
+            total,
+            pageSize <= 0 ? 0 : (int) Math.ceil((double) total / pageSize),
+            page < (pageSize <= 0 ? 0 : (int) Math.ceil((double) total / pageSize))
+        );
+    }
+}
